+++ src/Component/ContentElement/TabStartElementController.php
@@ -12,12 +12,16 @@ use Symfony\Component\HttpFoundation\Response;
 
 use function rtrim;
 
-/** @ContentElement("bs_tab_start", category="bootstrap_tabs") */
+/** @ContentElement("bs_tab_start", category="bootstrap_tabs", template="ce_bs_tab_start") */
 final class TabStartElementController extends AbstractTabElementController
 {
     /** {@inheritDoc} */
-    protected function preGenerate(Request $request, Model $model, string $section, ?array $classes = null): ?Response
-    {
+    protected function preGenerate(
+        Request $request,
+        Model $model,
+        string $section,
+        array|null $classes = null,
+    ): Response|null {
         if (! $this->isBackendRequest($request)) {
             return null;
         }
@@ -51,7 +55,7 @@ final class TabStartElementController extends AbstractTabElementController
         return $data;
     }
 
-    protected function getParent(ContentModel $model): ?ContentModel
+    protected function getParent(ContentModel $model): ContentModel|null
     {
         return $model;
     }