+++ src/ContaoManager/Plugin.php
@@ -17,7 +17,7 @@ use function class_exists;
 final class Plugin implements BundlePluginInterface
 {
     /**
-     * {@inheritdoc}
+     * {@inheritDoc}
      */
     public function getBundles(ParserInterface $parser): array
     {