+++ src/DependencyInjection/ContaoBootstrapTabExtension.php
@@ -9,19 +9,14 @@ use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Extension\Extension;
 use Symfony\Component\DependencyInjection\Loader\XmlFileLoader;
 
-use function dirname;
-
 final class ContaoBootstrapTabExtension extends Extension
 {
     /**
-     * {@inheritdoc}
+     * {@inheritDoc}
      */
     public function load(array $configs, ContainerBuilder $container): void
     {
-        $loader = new XmlFileLoader(
-            $container,
-            new FileLocator(dirname(__DIR__) . '/Resources/config')
-        );
+        $loader = new XmlFileLoader($container, new FileLocator(__DIR__ . '/../Resources/config'));
 
         $loader->load('config.xml');
         $loader->load('services.xml');