+++ src/EventListener/Dca/ContentListener.php
@@ -13,7 +13,7 @@ use Contao\StringUtil;
 use Netzmacht\Contao\Toolkit\Data\Model\RepositoryManager;
 use Netzmacht\Contao\Toolkit\Dca\DcaManager;
 use Netzmacht\Contao\Toolkit\Dca\Listener\AbstractListener;
-use Netzmacht\Contao\Toolkit\View\Assets\AssetsManager;
+use stdClass;
 
 use function array_unshift;
 use function assert;
@@ -22,20 +22,16 @@ use function time;
 
 final class ContentListener extends AbstractListener
 {
-    /**@var string */
-    // phpcs:ignore SlevomatCodingStandard.TypeHints.PropertyTypeHint.MissingNativeTypeHint
-    protected static $name = 'tl_content';
-
-    private RepositoryManager $repositories;
-
-    private AssetsManager $assetsManager;
-
-    public function __construct(DcaManager $dcaManager, RepositoryManager $repositories, AssetsManager $assetsManager)
-    {
+    public function __construct(
+        DcaManager $dcaManager,
+        private readonly RepositoryManager $repositories,
+    ) {
         parent::__construct($dcaManager);
+    }
 
-        $this->repositories  = $repositories;
-        $this->assetsManager = $assetsManager;
+    public static function getName(): string
+    {
+        return 'tl_content';
     }
 
     /**
@@ -45,15 +41,13 @@ final class ContentListener extends AbstractListener
      */
     public function initializeDca(): void
     {
-        $this->assetsManager->addStylesheet('contao_bootstrap_tab::css/backend.css');
-
         if (! $this->getDefinition()->has(['fields', 'bs_grid'])) {
             return;
         }
 
         $this->getDefinition()->modify(
             ['fields', 'bs_grid', 'load_callback'],
-            static function (?array $value): array {
+            static function (array|null $value): array {
                 $value   = (array) $value;
                 $value[] = [
                     'contao_bootstrap.tab.listener.dca.content',
@@ -61,7 +55,7 @@ final class ContentListener extends AbstractListener
                 ];
 
                 return $value;
-            }
+            },
         );
     }
 
@@ -73,7 +67,7 @@ final class ContentListener extends AbstractListener
      * @SuppressWarnings(PHPMD.Superglobals)
      * @Callback(table="tl_content", target="fields.bs_tab_parent.options")
      */
-    public function getTabParentOptions(): array
+    public function getTabParentOptions(DataContainer $dataContainer): array
     {
         $columns = [
             'tl_content.type = ?',
@@ -81,10 +75,9 @@ final class ContentListener extends AbstractListener
             'tl_content.ptable = ?',
         ];
 
-        /** @psalm-suppress UndefinedConstant */
         $values = [
             'bs_tab_start',
-            CURRENT_ID,
+            $dataContainer->currentPid,
             $this->getDefinition()->get(['config', 'ptable']),
         ];
 
@@ -96,7 +89,7 @@ final class ContentListener extends AbstractListener
                 $options[$model->id] = sprintf(
                     '%s [%s]',
                     $model->bs_tab_name,
-                    $model->id
+                    $model->id,
                 );
             }
         }
@@ -109,10 +102,8 @@ final class ContentListener extends AbstractListener
      *
      * @param mixed              $value         The field value.
      * @param DataContainer|null $dataContainer The data container driver.
-     *
-     * @return mixed
      */
-    public function configureGridField($value, ?DataContainer $dataContainer)
+    public function configureGridField(mixed $value, DataContainer|null $dataContainer): mixed
     {
         if ($dataContainer && $dataContainer->activeRecord && $dataContainer->activeRecord->type === 'bs_tab_start') {
             $this->getDefinition()->set(['fields', 'bs_grid', 'eval', 'mandatory'], false);
@@ -135,7 +126,7 @@ final class ContentListener extends AbstractListener
         }
 
         $current = $dataContainer->activeRecord;
-        assert($current instanceof Result || $current instanceof ContentModel);
+        assert($current instanceof Result || $current instanceof ContentModel || $current instanceof stdClass);
 
         $stopElement  = $this->getStopElement($current);
         $count        = $this->countRequiredSeparators($dataContainer->activeRecord->bs_tabs, $current);
@@ -155,10 +146,10 @@ final class ContentListener extends AbstractListener
     /**
      * Count the required separator fields which should be created.
      *
-     * @param mixed               $definition The tab definitions.
-     * @param ContentModel|Result $current    The current content model.
+     * @param mixed                        $definition The tab definitions.
+     * @param ContentModel|Result|stdClass $current    The current content model.
      */
-    private function countRequiredSeparators($definition, $current): int
+    private function countRequiredSeparators(mixed $definition, ContentModel|Result|stdClass $current): int
     {
         $definition = StringUtil::deserialize($definition, true);
         $count      = -1;
@@ -184,11 +175,11 @@ final class ContentListener extends AbstractListener
     /**
      * Create separators.
      *
-     * @param int                 $value   Number of separators being created.
-     * @param ContentModel|Result $current Current model.
-     * @param int                 $sorting Current sorting value.
+     * @param int                          $value   Number of separators being created.
+     * @param ContentModel|Result|stdClass $current Current model.
+     * @param int                          $sorting Current sorting value.
      */
-    protected function createSeparators(int $value, $current, int $sorting): int
+    protected function createSeparators(int $value, ContentModel|Result|stdClass $current, int $sorting): int
     {
         for ($count = 1; $count <= $value; $count++) {
             $sorting += 8;
@@ -221,10 +212,10 @@ final class ContentListener extends AbstractListener
     /**
      * Create the stop element.
      *
-     * @param ContentModel|Result $current Model.
-     * @param int                 $sorting Last sorting value.
+     * @param ContentModel|Result|stdClass $current Model.
+     * @param int                          $sorting Last sorting value.
      */
-    protected function createStopElement($current, int $sorting): ContentModel
+    protected function createStopElement(ContentModel|Result|stdClass $current, int $sorting): ContentModel
     {
         $sorting += 8;
 
@@ -234,12 +225,15 @@ final class ContentListener extends AbstractListener
     /**
      * Create a tab element.
      *
-     * @param ContentModel|Result $current Current content model.
-     * @param string              $type    Type of the content model.
-     * @param int                 $sorting The sorting value.
+     * @param ContentModel|Result|stdClass $current Current content model.
+     * @param string                       $type    Type of the content model.
+     * @param int                          $sorting The sorting value.
      */
-    protected function createTabElement($current, string $type, int &$sorting): ContentModel
-    {
+    protected function createTabElement(
+        ContentModel|Result|stdClass $current,
+        string $type,
+        int &$sorting,
+    ): ContentModel {
         $model                = new ContentModel();
         $model->tstamp        = time();
         $model->pid           = $current->pid;
@@ -255,11 +249,11 @@ final class ContentListener extends AbstractListener
     /**
      * Get the next content elements.
      *
-     * @param ContentModel|Result $current Current content model.
+     * @param ContentModel|Result|stdClass $current Current content model.
      *
      * @return ContentModel[]
      */
-    protected function getNextElements($current): array
+    protected function getNextElements(ContentModel|Result|stdClass $current): array
     {
         $collection = $this->repositories->getRepository(ContentModel::class)->findBy(
             [
@@ -268,7 +262,7 @@ final class ContentListener extends AbstractListener
                 'tl_content.sorting > ?',
             ],
             [$current->ptable, $current->pid, $current->sorting],
-            ['order' => 'tl_content.sorting ASC']
+            ['order' => 'tl_content.sorting ASC'],
         );
 
         if ($collection) {
@@ -281,13 +275,13 @@ final class ContentListener extends AbstractListener
     /**
      * Get related stop element.
      *
-     * @param ContentModel|Result $current Current element.
+     * @param ContentModel|Result|stdClass $current Current element.
      */
-    protected function getStopElement($current): ContentModel
+    protected function getStopElement(ContentModel|Result|stdClass $current): ContentModel
     {
         $stopElement = $this->repositories->getRepository(ContentModel::class)->findOneBy(
             ['tl_content.type=?', 'tl_content.bs_tab_parent=?'],
-            ['bs_tab_end', $current->id]
+            ['bs_tab_end', $current->id],
         );
 
         if ($stopElement instanceof ContentModel) {