+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 Changelog
 =========
 
+3.0.1 (2024-01-26)
+------------------
+
+### Added
+
+ - Add compatibility with Contao 5.3
+
+3.0.0 (2023-06-12)
+------------------
+
+[Full Changelog](https://github.com/contao-bootstrap/tab/compare/2.1.1...3.0.0)
+
+### Changed
+
+ - Use mvo/contao-group-widget insteadof multi column wizard
+ - Support Contao 5 and Bootstrap 5
+
 2.1.1 (2023-04-17)
 ------------------